Mer småplukk funnet under gjennomlesing av innvilgelsesbrev
