+++ pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/etterlatte/fixtures/OmstillingsstoenadRevurdering.kt
@@ -3,6 +3,7 @@ package no.nav.pensjon.etterlatte.fixtures
 import no.nav.pensjon.brevbaker.api.model.*
 import no.nav.pensjon.etterlatte.maler.*
 import no.nav.pensjon.etterlatte.maler.omstillingsstoenad.revurdering.OmstillingsstoenadRevurderingDTO
+import no.nav.pensjon.etterlatte.maler.omstillingsstoenad.revurdering.OmstillingsstoenadRevurderingRedigerbartUtfallDTO
 import java.time.LocalDate
 
 fun createOmstillingsstoenadRevurderingDTO() =
@@ -74,6 +75,13 @@ fun createOmstillingsstoenadRevurderingDTO() =
         datoVedtakOmgjoering = null,
         harFlereUtbetalingsperioder = true,
         lavEllerIngenInntekt = false,
+        harUtbetaling = true,
+        feilutbetaling = FeilutbetalingType.FEILUTBETALING_MED_VARSEL
+    )
+
+fun createOmstillingsstoenadRevurderingRedigerbartUtfallDTO() =
+    OmstillingsstoenadRevurderingRedigerbartUtfallDTO(
+        erEtterbetaling = true,
         harUtbetaling = true,
         feilutbetaling = FeilutbetalingType.FEILUTBETALING_MED_VARSEL
     )
