--- pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/etterlatte/fixtures/OmstillingsstoenadRevurderingRedigerbartUtfallDTO.kt
@@ -1,11 +1,0 @@
-package no.nav.pensjon.etterlatte.fixtures
-
-import no.nav.pensjon.etterlatte.maler.FeilutbetalingType
-import no.nav.pensjon.etterlatte.maler.omstillingsstoenad.revurdering.OmstillingsstoenadRevurderingRedigerbartUtfallDTO
-
-fun createOmstillingsstoenadRevurderingRedigerbartUtfallDTO() =
-    OmstillingsstoenadRevurderingRedigerbartUtfallDTO(
-        erEtterbetaling = true,
-        harUtbetaling = true,
-        feilutbetaling = FeilutbetalingType.FEILUTBETALING_MED_VARSEL
-    )
